workspaces/core/pyproject.toml
[tool.poetry]
name = "balpy-core"
version = "0.0.0"
description = ""
authors = ["José Ribeiro <jose@bleu.studio>"]
packages = [{include = "balpy", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
joblib = "^1.3.2"
web3 = "^6.8.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
